widen-system-uid: fix MNTR quarantine specs still expecting int uids
CI (build 128643) failed 5 multi-node tests in Akka.Remote.Tests.MultiNode:
the Subject actors reply with AddressUidExtension.Uid(...) - now long - but
the driver-side helpers still asserted ExpectMsg<int>/(int, IActorRef).
Tell(object) erases the static type, so this compiled clean and only failed
the runtime type check, on every run. Node2's port-bind failure in
RemoteQuarantinePiercingSpec was collateral from the aborted barrier flow.

Widened the ExpectMsg type params and uid locals to long in
RemoteRestartedQuarantinedSpec, RemoteQuarantinePiercingSpec, and
PiercingShouldKeepQuarantineSpec. No production change. All three specs
pass locally through the MultiNode test adapter (6/6, net10.0 Release).

+++ src/core/Akka.Remote.Tests.MultiNode/PiercingShouldKeepQuarantineSpec.cs
@@ -78,7 +78,7 @@ namespace Akka.Remote.Tests.MultiNode
 
                 // Communicate with second system
                 Sys.ActorSelection(Node(_config.Second) / "user" / "subject").Tell("getuid");
-                var uid = ExpectMsg<int>(TimeSpan.FromSeconds(10));
+                var uid = ExpectMsg<long>(TimeSpan.FromSeconds(10));
                 EnterBarrier("actor-identified");
 
                 // Manually Quarantine the other system
